Rename pallmall test app to H2MS
