+++ pom.xml
@@ -4,11 +4,11 @@
 	<modelVersion>4.0.0</modelVersion>
 
 	<groupId>edu.harvard</groupId>
-	<artifactId>pallmall</artifactId>
+	<artifactId>h2ms</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 	<packaging>jar</packaging>
 
-	<name>pallmall</name>
+	<name>h2ms</name>
 	<description>Demo project for Spring Boot</description>
 
 	<parent>
@@ -92,7 +92,7 @@
 			<artifactId>mail</artifactId>
 			<version>1.4</version>
 		</dependency>
-		
+
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-security</artifactId>
