+++ src/main/java/edu/harvard/h2ms/H2MSRestAppInitializer.java
@@ -1,4 +1,4 @@
-package edu.harvard.pallmall;
+package edu.harvard.h2ms;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
