+++ src/main/java/edu/harvard/h2ms/repository/DoctorRepository.java
@@ -1,11 +1,12 @@
-package edu.harvard.pallmall.repository;
+package edu.harvard.h2ms.repository;
 
 import java.util.List;
 
-import edu.harvard.pallmall.domain.core.Doctor;
 import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.repository.query.Param;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
+
+import edu.harvard.h2ms.domain.core.Doctor;
 
 @RepositoryRestResource(collectionResourceRel = "doctors", path = "doctors")
 public interface DoctorRepository extends PagingAndSortingRepository<Doctor, Long> {
