+++ src/main/java/edu/harvard/h2ms/repository/EventRepository.java
@@ -1,7 +1,8 @@
-package edu.harvard.pallmall.repository;
+package edu.harvard.h2ms.repository;
 
-import edu.harvard.pallmall.domain.core.Event;
 import org.springframework.data.repository.PagingAndSortingRepository;
+
+import edu.harvard.h2ms.domain.core.Event;
 
 public interface EventRepository extends PagingAndSortingRepository<Event, Long> {
