+++ src/main/java/edu/harvard/h2ms/repository/WristBandRepository.java
@@ -1,7 +1,8 @@
-package edu.harvard.pallmall.repository;
+package edu.harvard.h2ms.repository;
 
-import edu.harvard.pallmall.domain.core.WristBand;
 import org.springframework.data.repository.PagingAndSortingRepository;
+
+import edu.harvard.h2ms.domain.core.WristBand;
 
 public interface WristBandRepository extends PagingAndSortingRepository<WristBand, Long> {
