+++ src/main/java/edu/harvard/h2ms/web/controller/ManagementDashboardController.java
@@ -1,11 +1,12 @@
-package edu.harvard.pallmall.web.controller;
+package edu.harvard.h2ms.web.controller;
 
-import edu.harvard.pallmall.domain.core.Event;
-import edu.harvard.pallmall.service.ManagementDashboardService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
+
+import edu.harvard.h2ms.domain.core.Event;
+import edu.harvard.h2ms.service.ManagementDashboardService;
 
 import java.sql.Time;
